refactor(profile): profile 페이지 및 컴포넌트 폴더구조 변경
- page 폴더내 profile 파일만 남기고 나머지는 컴포넌트 폴더로 이동

#36
