+++ src/pages/Profile.tsx
@@ -1,0 +1,14 @@
+import User from '../components/Profile/User';
+import Pet from '../components/Profile/Pet';
+import EditPet from '../components/Profile/EditPet';
+
+function Profile() {
+  return (
+    <div>
+      <User />
+    </div>
+  );
+}
+
+export default Profile;
+export { Pet, EditPet };
