--- src/pages/Profile/index.tsx
@@ -1,14 +1,0 @@
-import User from './User';
-import Pet from './Pet';
-import EditPet from './EditPet';
-
-function Profile() {
-  return (
-    <div>
-      <User />
-    </div>
-  );
-}
-
-export default Profile;
-export { Pet, EditPet };
